Worked around InheritedResource for login

+++ app/controllers/user_session_controller.rb
@@ -5,22 +5,21 @@ class UserSessionController < InheritedResources::Base
   defaults :singleton => true
 
   def create
-    create! do |success, failure|
-      success.html do
-        # somehow current_user is sometimes nil here --AB
-        unless current_user.nil?
-          if current_user.disabled?
-            flash[:error] = _("Your account has been disabled, please contact support.")
-            flash.delete(:notice)
-            current_user_session.destroy
-            redirect_to login_path
-          else
-            redirect_to home_path
-          end
+    sess = UserSession.create(email: params[:user_session][:email], password: params[:user_session][:password], remember_me: params[:user_session][:remember_me])
+    if sess.valid?
+      sess.save
+      unless current_user.nil?
+        if current_user.disabled?
+          flash[:error] = _("Your account has been disabled, please contact support.")
+          flash.delete(:notice)
+          current_user_session.destroy
+          redirect_to login_path
         else
-          current_user_session.destroy unless current_user_session.nil?
           redirect_to home_path
         end
+      else
+        current_user_session.destroy unless current_user_session.nil?
+        redirect_to home_path
       end
     end
   end
